Fix: change type to enum for languages
Changed `EditorLanguage` from a type to an enum. It was very complicated to find all the instances and values of this type, using an enum and 2 helper functions it allows for all of this to be totally centralized. It will make the code cleaner, more easily maintainable and it will help add new languages effortlessly. I let the tag system on the module like how it is now, maybe it is also something to improve upon ?